requirements.txt
PyJWT>=2.0
Flask
flask-restful
flask-cors
python-dotenv
flask_sqlalchemy
werkzeug
pytest
requests